Added functionality to the player control
The player control can now has a FSM for animation attached. While no
animationa are currently added. It will blend between the defeated and
standard animations for walking and idles. Mining animation shouls play,
knockdown animation can be called by a function.

Mining is based on the animation. Pressing E at a certain point in the
animation will cause the player character to mine. The values will be
determined when we have the animation but will be based on the apex of
the swing.

Pressing E while facing a game object with the Minable Layer will cause
the player to enter the mining state.